@@ -12,48 +12,59 @@
   },
   "dependencies": {
     "@expo/samples": "2.1.1",
+    "@react-native-community/masked-view": "0.1.10",
+    "@react-navigation/bottom-tabs": "^5.9.2",
+    "@react-navigation/drawer": "^5.9.3",
+    "@react-navigation/material-top-tabs": "^5.3.9",
+    "@react-navigation/native": "^5.8.10",
+    "@react-navigation/stack": "^5.9.3",
     "axios": "^0.19.0",
     "color": "^3.1.1",
-    "expo": "^38.0.2",
-    "expo-analytics-amplitude": "~8.2.1",
-    "expo-asset": "~8.1.7",
-    "expo-constants": "~9.1.1",
-    "expo-device": "~2.2.1",
-    "expo-font": "~8.2.1",
-    "expo-haptics": "~8.2.1",
+    "expo": "^39.0.0",
+    "expo-analytics-amplitude": "~8.3.1",
+    "expo-asset": "~8.2.0",
+    "expo-branch": "~2.3.1",
+    "expo-constants": "~9.2.0",
+    "expo-device": "~2.3.0",
+    "expo-font": "~8.3.0",
+    "expo-haptics": "~8.3.0",
+    "expo-intent-launcher": "~8.3.0",
+    "expo-linear-gradient": "~8.3.0",
     "expo-linking": "^1.0.1",
-    "expo-splash-screen": "~0.6.1",
-    "expo-status-bar": "^1.0.0",
-    "expo-web-browser": "~8.3.1",
+    "expo-notifications": "~0.7.2",
+    "expo-splash-screen": "~0.6.2",
+    "expo-status-bar": "~1.0.2",
+    "expo-web-browser": "~8.5.0",
     "geolib": "^3.0.3",
+    "html-entities": "^1.3.1",
     "lodash": "^4.17.15",
     "lottie-react-native": "~2.6.1",
     "moment": "^2.24.0",
     "normalizr": "^3.3.0",
-    "react": "16.11.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.2.tar.gz",
-    "react-native-gesture-handler": "^1.6.1",
+    "react": "16.13.1",
+    "react-devtools": "^4.10.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.3.tar.gz",
+    "react-native-gesture-handler": "~1.7.0",
     "react-native-paper": "^4.2.0",
-    "react-native-reanimated": "^1.9.0",
+    "react-native-reanimated": "~1.13.0",
     "react-native-render-html": "^4.2.0",
-    "react-native-screens": "^2.11.0",
+    "react-native-safe-area-context": "3.1.4",
+    "react-native-screens": "~2.10.1",
     "react-native-slider": "^0.11.0",
     "react-native-snap-carousel": "^3.7.5",
-    "react-native-webview": "9.4.0",
-    "react-navigation": "^3.11.1",
-    "react-navigation-header-buttons": "^2.2.0",
+    "react-native-tab-view": "^2.15.2",
+    "react-native-webview": "10.7.0",
     "react-redux": "^6.0.1",
     "redux": "^4.0.1",
     "redux-persist": "^5.10.0",
     "redux-saga": "^1.0.2",
     "redux-sentry-middleware": "0.0.15",
     "sentry-expo": "^2.1.2",
-    "ws": "^7.0.0",
-    "expo-branch": "~2.2.1"
+    "ws": "^7.0.0"
   },
   "devDependencies": {
-    "babel-preset-expo": "^8.2.3",
-    "jest-expo": "^38.0.0",
+    "babel-preset-expo": "^8.3.0",
+    "jest-expo": "^39.0.0",
     "redux-logger": "^3.0.6"
   },
   "private": true
