@@ -1,4 +1,5 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
+import Color from 'color'
 import {
     View,
     FlatList,
@@ -9,15 +10,20 @@ import {
     TouchableOpacity,
     Platform,
 } from 'react-native'
+import { connect } from 'react-redux'
 import Moment from 'moment'
-import Color from 'color'
+
 import HTML from 'react-native-render-html'
 
 import { getActiveDomain } from '../redux/domains'
 import { handleArticlePress } from '../utils/articlePress'
-import { asyncFetchFeaturedImage, asyncFetchComments } from '../utils/sagaHelpers'
+import { asyncFetchArticle } from '../utils/sagaHelpers'
 
-import { connect } from 'react-redux'
+import ArticleAuthors from '../components/Article/ArticleAuthors'
+
+import { Html5Entities } from 'html-entities'
+
+const entities = new Html5Entities()
 
 Moment.updateLocale('en', {
     relativeTime: {
@@ -25,210 +31,158 @@ Moment.updateLocale('en', {
     },
 })
 
-class RelatedStoriesList extends React.Component {
-    state = {
-        stories: [],
-        error: '',
-    }
+const RelatedStoriesWidget = ({ relatedStoryIds, activeDomain, theme }) => {
+    const [stories, setStories] = useState([])
+    const [error, setError] = useState('')
 
-    componentDidMount() {
-        if (this.props.relatedStoryIds) {
-            this._fetchRelatedStories(this.props.relatedStoryIds)
+    useEffect(() => {
+        if (relatedStoryIds?.length) {
+            _fetchRelatedStories(relatedStoryIds)
         }
-    }
+    }, [relatedStoryIds])
 
-    render() {
-        return (
-            <View
-                style={{
-                    flex: 1,
-                    justifyContent: 'center',
-                    padding: 10,
-                    borderRadius: 8,
-                }}
-            >
-                <Text
-                    style={{
-                        textAlign: 'center',
-                        color: '#757575',
-                        fontSize: 19,
-                        fontWeight: 'bold',
-                        paddingVertical: 10,
-                    }}
-                >
-                    Related Stories
-                </Text>
-                {this._renderRelatedStories()}
-            </View>
-        )
-    }
-
-    _renderRelatedStories = () => {
-        const { relatedStoryIds } = this.props
-        const { stories, error } = this.state
-        if (error) {
-            return <Text style={{ textAlign: 'center' }}>Error loading related stories</Text>
-        } else if (relatedStoryIds && stories.length == 0) {
-            return <ActivityIndicator />
-        } else {
-            return (
-                <View style={{ flex: 1 }}>
-                    <FlatList
-                        // style={{ flex: 1, marginVertical: 5 }}
-                        data={stories}
-                        keyExtractor={(item) => item.id.toString()}
-                        renderItem={this._renderItem}
-                    />
-                </View>
-            )
-        }
-    }
-
-    _renderItem = ({ item }) => {
-        const { theme, activeDomain } = this.props
-        const article = item
-        return (
-            <TouchableOpacity
-                style={{ flex: 1 }}
-                onPress={() => handleArticlePress(article, activeDomain, true)}
-            >
-                <View style={styles.storyContainer}>
-                    {article.featuredImage ? (
-                        <Image
-                            source={{ uri: article.featuredImage.uri }}
-                            style={styles.featuredImage}
-                        />
-                    ) : null}
-                    <View style={styles.storyInfo}>
-                        <HTML
-                            html={article.title.rendered}
-                            baseFontStyle={{ fontSize: 16 }}
-                            customWrapper={(text) => {
-                                return (
-                                    <Text ellipsizeMode='tail' numberOfLines={2}>
-                                        {text}
-                                    </Text>
-                                )
-                            }}
-                            tagsStyles={{
-                                rawtext: {
-                                    fontSize: 16,
-                                    fontWeight: 'bold',
-                                    color: theme.dark ? 'white' : 'black',
-                                },
-                            }}
-                        />
-                        <Text
-                            ellipsizeMode='tail'
-                            numberOfLines={1}
-                            style={{
-                                color: theme.colors.accent,
-                                fontSize: 14,
-                            }}
-                        >
-                            {article.custom_fields.writer
-                                ? this._renderWriters(article.custom_fields.writer)
-                                : ''}
-                        </Text>
-                        <View
-                            style={{
-                                flexDirection: 'row',
-                                alignItems: 'center',
-                                justifyContent: 'space-between',
-                            }}
-                        >
-                            <View style={{ flexDirection: 'row' }}>
-                                {this._renderDate(article.date)}
-                            </View>
-                        </View>
-                    </View>
-                </View>
-            </TouchableOpacity>
-        )
-    }
-
-    _renderDate = (date) => {
-        return (
-            <Text
-                style={{
-                    fontSize: 14,
-                    color: '#9e9e9e',
-                }}
-            >
-                {Moment().isAfter(Moment(date).add(7, 'days'))
-                    ? String(Moment(date).format('MMM D, YYYY'))
-                    : String(Moment(date).fromNow())}
-            </Text>
-        )
-    }
-
-    _renderWriters = (writers) => {
-        let newArr = ''
-        for (let i = 0; i < writers.length; i++) {
-            if (i === writers.length - 2) {
-                newArr += `${writers[i]} & `
-            } else if (i === writers.length - 1) {
-                newArr += `${writers[i]}`
-            } else {
-                newArr += `${writers[i]}, `
-            }
-        }
-        return newArr
-    }
-
-    _getStory = async (id) => {
-        const { activeDomain } = this.props
-        const result = await fetch(`https://${activeDomain.url}/wp-json/wp/v2/posts/${id}`)
-        const post = await result.json()
-        if (post._links['wp:featuredmedia']) {
-            await asyncFetchFeaturedImage(`${post._links['wp:featuredmedia'][0].href}`, post)
-        }
-        await asyncFetchComments(activeDomain.url, post)
-        return post
-    }
-
-    _fetchRelatedStories = async (storyIds) => {
+    const _fetchRelatedStories = async (storyIds) => {
         try {
             const stories = await Promise.all(
                 storyIds.map(async (id) => {
-                    return await this._getStory(Number(id))
+                    return await asyncFetchArticle(activeDomain.url, Number(id))
                 })
             )
-            this.setState({
-                stories,
-            })
+            setStories(stories)
         } catch (err) {
             console.log('error fetching related stories', err)
-            this.setState({
-                error: true,
-            })
+            setError(true)
         }
     }
-}
 
-const styles = StyleSheet.create({
-    storyContainer: {
-        flexDirection: 'row',
-        flex: 1,
-        marginHorizontal: 10,
-        marginVertical: 10,
-        padding: 10,
-    },
-    featuredImage: {
-        width: 95,
-        height: 60,
-        borderRadius: 8,
-    },
-    storyInfo: {
-        flex: 1,
-        marginLeft: 10,
-        justifyContent: 'space-between',
-    },
-})
+    const _renderDate = (date) => {
+        if (Moment().isAfter(Moment(date).add(7, 'days'))) return Moment(date).format('MMM D, YYYY')
+        return String(Moment(date).fromNow())
+    }
+
+    const _renderRelatedStories = () => {
+        if (error) {
+            return (
+                <Text style={{ textAlign: 'center', color: theme.colors.text }}>
+                    Error loading related stories
+                </Text>
+            )
+        } else if (relatedStoryIds && !stories.length) {
+            return <ActivityIndicator />
+        } else {
+            return (
+                <View
+                    style={{
+                        flex: 1,
+                        marginLeft: -20,
+                        backgroundColor: Color(theme.colors.text).grayscale().alpha(0.05),
+                    }}
+                >
+                    {stories.map((story) => {
+                        return (
+                            <TouchableOpacity
+                                key={story.id}
+                                style={{ flex: 1 }}
+                                onPress={() => handleArticlePress(story, activeDomain)}
+                            >
+                                <View
+                                    style={{
+                                        flexDirection: 'row',
+                                        flex: 1,
+                                        margin: 10,
+                                        padding: 10,
+                                    }}
+                                >
+                                    {story.featuredImage ? (
+                                        <Image
+                                            source={{ uri: story.featuredImage.uri }}
+                                            style={{
+                                                width: 95,
+                                                height: 60,
+                                                marginRight: 10,
+                                                borderRadius: 8,
+                                            }}
+                                        />
+                                    ) : null}
+                                    <View
+                                        style={{
+                                            flex: 1,
+                                            justifyContent: 'space-between',
+                                        }}
+                                    >
+                                        <Text
+                                            style={{
+                                                fontSize: 16,
+                                                fontWeight: 'bold',
+                                                color: theme.colors.text,
+                                            }}
+                                            ellipsizeMode='tail'
+                                            numberOfLines={2}
+                                        >
+                                            {entities.decode(story.title?.rendered || '')}
+                                        </Text>
+                                        <View
+                                            style={{
+                                                alignItems: 'flex-start',
+                                            }}
+                                        >
+                                            <ArticleAuthors article={story} theme={theme} />
+                                        </View>
+                                        <View
+                                            style={{
+                                                marginTop: 10,
+                                                justifyContent: 'space-between',
+                                            }}
+                                        >
+                                            <View style={{ flexDirection: 'row' }}>
+                                                <Text
+                                                    style={{
+                                                        fontSize: 14,
+                                                        color: theme.colors.grayText,
+                                                    }}
+                                                >
+                                                    {_renderDate(story.date)}
+                                                </Text>
+                                            </View>
+                                        </View>
+                                    </View>
+                                </View>
+                            </TouchableOpacity>
+                        )
+                    })}
+                </View>
+            )
+        }
+    }
+
+    return (
+        <View
+            style={{
+                flex: 1,
+                justifyContent: 'center',
+                padding: 10,
+                borderRadius: 8,
+            }}
+        >
+            <Text
+                style={{
+                    textAlign: 'center',
+                    color: theme.colors.grayText,
+                    fontSize: 19,
+                    fontFamily: 'ralewayBold',
+                    paddingVertical: 10,
+                }}
+            >
+                Related Stories
+            </Text>
+            {_renderRelatedStories()}
+        </View>
+    )
+}
 
 const mapStateToProps = (state) => ({
     activeDomain: getActiveDomain(state),
     theme: state.theme,
 })
 
-export default connect(mapStateToProps)(RelatedStoriesList)
+export default connect(mapStateToProps)(RelatedStoriesWidget)
