@@ -2,19 +2,24 @@ export const types = {
     FETCH_PROFILES: 'FETCH_PROFILES',
     REQUEST_PROFILES: 'REQUEST_PROFILES',
     RECEIVE_PROFILES: 'RECEIVE_PROFILES',
+    FETCH_PROFILE: 'FETCH_PROFILE',
+    FETCH_PROFILE_REQUEST: 'FETCH_PROFILE_REQUEST',
+    FETCH_PROFILE_SUCCESS: 'FETCH_PROFILE_SUCCESS',
+    FETCH_PROFILE_ERROR: 'FETCH_PROFILE_ERROR',
     FETCH_PROFILE_ARTICLES: 'FETCH_PROFILE_ARTICLES',
     SET_PROFILE_ARTICLES: 'SET_PROFILE_ARTICLES',
     CLEAR_PROFILE_ARTICLES: 'CLEAR_PROFILE_ARTICLES',
     CLEAR_PROFILE_ERROR: 'CLEAR_PROFILE_ERROR',
-    SET_PROFILE_ARTICLE_ERROR: 'SET_PROFILE_ARTICLE_ERROR'
+    SET_PROFILE_ARTICLE_ERROR: 'SET_PROFILE_ARTICLE_ERROR',
 }
 
 export default function profiles(
     state = {
         isLoaded: false,
         items: [],
+        single: null,
         articles: [],
-        error: ''
+        error: '',
     },
     action
 ) {
@@ -22,33 +27,43 @@ export default function profiles(
         case types.REQUEST_PROFILES:
             return {
                 ...state,
-                isLoaded: false
+                isLoaded: false,
             }
         case types.RECEIVE_PROFILES:
             return {
                 ...state,
                 isLoaded: true,
-                items: action.profiles
+                items: action.profiles,
+            }
+        case types.FETCH_PROFILE_SUCCESS:
+            return {
+                ...state,
+                single: action.profile,
+            }
+        case types.FETCH_PROFILE_ERROR:
+            return {
+                ...state,
+                single: null,
             }
         case types.SET_PROFILE_ARTICLES:
             return {
                 ...state,
-                articles: action.payload
+                articles: action.payload,
             }
         case types.CLEAR_PROFILE_ARTICLES:
             return {
                 ...state,
-                articles: []
+                articles: [],
             }
         case types.SET_PROFILE_ARTICLE_ERROR:
             return {
                 ...state,
-                error: action.payload
+                error: action.payload,
             }
         case types.CLEAR_PROFILE_ERROR:
             return {
                 ...state,
-                error: ''
+                error: '',
             }
         default:
             return state
@@ -60,47 +75,70 @@ export const actions = {
         return {
             type: types.FETCH_PROFILES,
             domain,
-            year
+            year,
         }
     },
     requestProfiles() {
         return {
-            type: types.REQUEST_PROFILES
+            type: types.REQUEST_PROFILES,
         }
     },
     receiveProfiles(profiles) {
         return {
             type: types.RECEIVE_PROFILES,
-            profiles
+            profiles,
+        }
+    },
+    fetchProfile(profileId) {
+        return {
+            type: types.FETCH_PROFILE,
+            profileId,
+        }
+    },
+    fetchProfileRequest() {
+        return {
+            type: types.FETCH_PROFILE_REQUEST,
+        }
+    },
+    fetchProfileSuccess(profile) {
+        return {
+            type: types.FETCH_PROFILE_SUCCESS,
+            profile,
+        }
+    },
+    fetchProfileError(error) {
+        return {
+            type: types.FETCH_PROFILE_ERROR,
+            error,
         }
     },
     fetchProfileArticles(url, writerTermId) {
         return {
             type: types.FETCH_PROFILE_ARTICLES,
             url,
-            writerTermId
+            writerTermId,
         }
     },
     setProfileArticles(payload) {
         return {
             type: types.SET_PROFILE_ARTICLES,
-            payload
+            payload,
         }
     },
     clearProfileArticles() {
         return {
-            type: types.CLEAR_PROFILE_ARTICLES
+            type: types.CLEAR_PROFILE_ARTICLES,
         }
     },
     setProfileArticleError(payload) {
         return {
             type: types.SET_PROFILE_ARTICLE_ERROR,
-            payload
+            payload,
         }
     },
     clearProfileError() {
         return {
-            type: types.CLEAR_PROFILE_ERROR
+            type: types.CLEAR_PROFILE_ERROR,
         }
-    }
+    },
 }
