@@ -1,12 +1,14 @@
 import { put, call, takeLatest, all, select, fork } from 'redux-saga/effects'
 
 import { types as globalTypes, actions as globalActions } from '../redux/global'
+import { actions as domainsActions } from '../redux/domains'
 import { types as adTypes, actions as adActions } from '../redux/ads'
 import { actions as userActions, getApiToken, getSubscribeAll, getFromPush } from '../redux/user'
 import { actions as themeActions } from '../redux/theme'
 import { actions as articleActions } from '../redux/articles'
 import { actions as savedArticleActions } from '../redux/savedArticles'
-import { getActiveDomain } from '../redux/domains'
+import { actions as likedArticleActions } from '../redux/likedArticles'
+import { getActiveDomain, getSavedDomains } from '../redux/domains'
 
 import { fetchMenu } from '../sagas/menu'
 import {
@@ -14,23 +16,36 @@ import {
     subscribe,
     fetchNotificationSubscriptions,
     findOrCreateUser,
+    fetchUnreadStories,
 } from '../sagas/user'
-import { loadActiveDomain } from '../sagas/global'
 
 import domainApiService from '../api/domain'
 import apiService from '../api/api'
 
 import * as Linking from 'expo-linking'
 
-import NavigationService from '../utils/NavigationService'
 import { handleArticlePress } from '../utils/articlePress'
 
-import { SplashScreen } from 'expo'
 import * as Amplitude from 'expo-analytics-amplitude'
 import * as Sentry from 'sentry-expo'
-import Constants from 'expo-constants'
+import * as SplashScreen from 'expo-splash-screen'
 
-// get new home screen options
+async function hideSplashScreen() {
+    await SplashScreen.hideAsync()
+}
+
+function* initializeUser() {
+    try {
+        yield put(globalActions.initializeUserRequest())
+
+        yield call(findOrCreateUser)
+
+        yield put(globalActions.initializeUserSuccess())
+    } catch (err) {
+        console.log('error initializing user in saga', err)
+        yield put(globalActions.initializeUserError('error initializing user'))
+    }
+}
 
 function* startup(action) {
     // if coming from deep link need to handle that
@@ -40,16 +55,22 @@ function* startup(action) {
     const userSubscribeAll = yield select(getSubscribeAll)
     const fromPush = yield select(getFromPush)
     try {
+        yield put(globalActions.startupRequest())
+
         // set user domain for analytics
         Amplitude.setUserProperties({
             activeDomain: domain.id,
         })
 
-        yield put(globalActions.startupRequest())
         // get splash image right away
         const splashScreenUrl = yield call(getSplashScreenImage, domain)
         yield put(globalActions.receiveSplash(splashScreenUrl))
-        SplashScreen.hide()
+
+        try {
+            yield call(SplashScreen.hideAsync)
+        } catch (e) {
+            console.warn(e)
+        }
 
         // get menus and sync with DB -- save updated DB categories to push notification categories -- return obj with menu and DB categories
         const { menu, dbCategories } = yield call(fetchMenu, {
@@ -77,47 +98,51 @@ function* startup(action) {
         // get users notification subscriptions
         yield call(fetchNotificationSubscriptions, domain.id)
 
+        // get unread story ID's for user
+        yield call(fetchUnreadStories)
+
         //get domains custom options
         yield call(getCustomOptions, domain)
 
-        yield call(getHomeScreenArticles)
-
-        let mainCategory = menu[0].object_id
+        let activeCategory = menu[0].object_id
 
         yield put(
             articleActions.fetchArticlesIfNeeded({
                 domain: domain.url,
-                category: mainCategory,
+                category: activeCategory,
             })
         )
 
-        yield put(globalActions.setActiveCategory(mainCategory))
+        yield put(globalActions.setActiveCategory(activeCategory))
 
         yield put(savedArticleActions.initializeSaved(domain.id))
+        yield put(likedArticleActions.initializeLiked(domain.id))
 
-        if (fromPush) {
-            // go to main app
-            // NavigationService.navigate('MainApp')
-            NavigationService.nestedNavigate('MainApp', 'RecentStack')
+        // throw new Error()
 
-            yield call(handleArticlePress, fromPush, domain)
-            // yield call(handleArticlePress, testFromPush, domain)
-            // reset push key
-            yield put(userActions.setFromPush(false))
-        } else {
-            NavigationService.navigate('MainApp')
-        }
+        // if (fromPush) {
+        //     // go to main app
+        //     // NavigationService.navigate('MainApp')
+        //     NavigationService.nestedNavigate('MainApp', 'RecentStack')
+
+        //     yield call(handleArticlePress, fromPush, domain)
+        //     // yield call(handleArticlePress, testFromPush, domain)
+        //     // reset push key
+        //     yield put(userActions.setFromPush(false))
+        // } else {
+        //     NavigationService.navigate('MainApp')
+        // }
 
         yield put(globalActions.startupSuccess())
     } catch (err) {
-        console.log('initilize err', err)
+        console.log('startup saga error', err)
         // clear from push data if any is there
         yield put(userActions.setFromPush(false))
 
         // check if domain is still in DB
         const domainCheck = yield call(checkIfDomainIsInDb, domain.id)
 
-        if (domainCheck.length > 0) {
+        if (domainCheck.length) {
             yield put(globalActions.startupError('error initializing app'))
             Sentry.captureException(err)
         } else {
@@ -141,6 +166,7 @@ function* getHomeScreenArticles() {
                         articleActions.fetchArticlesIfNeeded({
                             domain: domain.url,
                             category: category,
+                            force: true,
                         })
                     )
                 })
@@ -155,175 +181,78 @@ function* getHomeScreenArticles() {
 
 function* getCustomOptions(domain) {
     try {
-        //get home page options
-        const [
-            headerImage,
-            headerLogo,
-            theme,
-            primary,
-            accent,
-            comments,
-            listStyle,
-            appAdOptions,
-        ] = yield all([
-            call(domainApiService.getCustomHeader, domain.url),
-            call(domainApiService.getCustomHeaderLogo, domain.url),
-            call(domainApiService.getCustomTheme, domain.url),
-            call(domainApiService.getCustomPrimaryColor, domain.url),
-            call(domainApiService.getCustomAccentColor, domain.url),
-            call(domainApiService.getCommentsToggle, domain.url),
-            call(domainApiService.getStoryListStyle, domain.url),
-            call(domainApiService.getAdOptions, domain.url),
-        ])
-
-        //home screen categories
-        const [
-            category1,
-            category2,
-            category3,
-            category1Amount,
-            category2Amount,
-            category3Amount,
-            homeScreenListStyle,
-            homeScreenMode,
-            homeScreenCategoryColor,
-        ] = yield all([
-            call(domainApiService.getHomeScreenCategory, domain.url, 1),
-            call(domainApiService.getHomeScreenCategory, domain.url, 2),
-            call(domainApiService.getHomeScreenCategory, domain.url, 3),
-            call(domainApiService.getHomeScreenCategoryAmount, domain.url, 1),
-            call(domainApiService.getHomeScreenCategoryAmount, domain.url, 2),
-            call(domainApiService.getHomeScreenCategoryAmount, domain.url, 3),
-            call(domainApiService.getHomeScreenListStyle, domain.url),
-            call(domainApiService.getHomeScreenMode, domain.url),
-            call(domainApiService.getCustomHomeCategoryColor, domain.url),
-        ])
-
-        if (homeScreenMode.result == 1 || homeScreenMode.result == '1') {
-            yield put(globalActions.receiveHomeScreenMode('legacy'))
-        } else {
-            yield put(globalActions.receiveHomeScreenMode('categories'))
-        }
-
-        const homeScreenCategories = []
-
-        if (!category1.result) {
-        } else {
-            homeScreenCategories.push(Number(category1.result))
-        }
-        if (!category2.result) {
-        } else {
-            homeScreenCategories.push(Number(category2.result))
-        }
-        if (!category3.result) {
-        } else {
-            homeScreenCategories.push(Number(category3.result))
-        }
-
-        const homeScreenCategoryAmounts = []
-
-        if (!category1Amount.result) {
-            homeScreenCategoryAmounts.push(5)
-        } else {
-            homeScreenCategoryAmounts.push(Number(category1Amount.result))
-        }
-        if (!category2Amount.result) {
-            homeScreenCategoryAmounts.push(5)
-        } else {
-            homeScreenCategoryAmounts.push(Number(category2Amount.result))
-        }
-        if (!category3Amount.result) {
-            homeScreenCategoryAmounts.push(5)
-        } else {
-            homeScreenCategoryAmounts.push(Number(category3Amount.result))
-        }
-
-        if (!homeScreenCategoryColor.result) {
-            homeScreenCategoryColor.result = null
-        }
-
-        yield put(globalActions.receiveAppAdOptions(appAdOptions))
-
-        yield fork(fetchAds, domain, appAdOptions)
-
-        yield put(globalActions.receiveHomeScreenCategoryColor(homeScreenCategoryColor.result))
-
-        yield put(globalActions.receiveHomeScreenCategories(homeScreenCategories))
-        yield put(globalActions.receiveHomeScreenCategoryAmounts(homeScreenCategoryAmounts))
-
-        yield put(
-            globalActions.receiveHomeScreenListStyle(
-                homeScreenListStyle.result === 'small' ||
-                    homeScreenListStyle.result === 'large' ||
-                    homeScreenListStyle.result === 'mix' ||
-                    homeScreenListStyle.result === 'alternating'
-                    ? homeScreenListStyle.result
-                    : 'small'
-            )
-        )
-
-        if (!theme.result) {
-            theme.result = 'light'
-        }
-        if (!primary.result) {
-            primary.result = '#2099CE'
-        }
-        if (!accent.result) {
-            accent.result = '#83B33B'
-        }
-
+        const results = yield call(domainApiService.getCustomOptions, domain.url)
+        // const categories = yield call(domainApiService.fetchCategories, domain.url)
+        console.log('results', results)
+        yield put(globalActions.receiveHeader(results.nav_header))
+        yield put(globalActions.receiveHeaderLogo(results.header_logo))
         yield put(
             themeActions.saveTheme({
-                theme: theme.result,
-                primary: primary.result,
-                accent: accent.result,
+                primary: results.primary_color,
+                accent: results.accent_color,
+                homeCategoryColor: results.home_category_color,
             })
         )
-        // // save if images are set otherwise empty string
-        yield put(
-            globalActions.receiveHeader(
-                headerImage[0] && headerImage[0].source_url ? headerImage[0].source_url : ''
-            )
-        )
-        yield put(
-            globalActions.receiveHeaderLogo(
-                headerLogo[0] && headerLogo[0].source_url ? headerLogo[0].source_url : ''
-            )
-        )
-        yield put(globalActions.receiveCommentsOption(comments.result === 'Enable' ? true : false))
-
+        const themeIsDark = yield select((state) => state.theme.dark)
+        if (themeIsDark) {
+            yield put(themeActions.toggleDarkMode(true))
+        }
+        yield put(globalActions.receiveCommentsOption(results.comments === 'Enable' ? true : false))
         yield put(
             globalActions.receiveStoryListStyle(
-                listStyle.result === 'small' ||
-                    listStyle.result === 'large' ||
-                    listStyle.result === 'mix' ||
-                    listStyle.result === 'alternating'
-                    ? listStyle.result
+                results.list_type === 'small' ||
+                    results.list_type === 'large' ||
+                    results.list_type === 'mix' ||
+                    results.list_type === 'alternating'
+                    ? results.list_type
                     : 'small'
             )
         )
+        yield put(globalActions.receiveAppAdOptions(results.ads))
+        yield put(globalActions.receiveHomeScreenMode(results.default_home))
+        yield put(
+            globalActions.receiveHomeScreenListStyle(
+                results.home_list_type === 'small' ||
+                    results.home_list_type === 'large' ||
+                    results.home_list_type === 'mix' ||
+                    results.home_list_type === 'alternating'
+                    ? results.home_list_type
+                    : 'small'
+            )
+        )
+        yield put(globalActions.receiveSportCenterOption(results.has_sportcenter))
 
-        try {
-            // save sportcenter option
-            const sportCenter = yield call(domainApiService.getSportCenterOption, domain.url)
-            yield put(globalActions.receiveSportCenterOption(sportCenter))
-        } catch (err) {
-            yield put(globalActions.receiveSportCenterOption(false))
+        yield fork(fetchAds, domain, results.ads)
+
+        // always fetch
+        if (true || results.legacy_home === 'categories') {
+            const results = yield call(domainApiService.getHomeScreenCategories, domain.url)
+
+            const homeScreenCategories = [
+                results.home_category_1,
+                results.home_category_2,
+                results.home_category_3,
+            ]
+            const homeScreenCategoryAmounts = [
+                results.home_category_1_amount,
+                results.home_category_2_amount,
+                results.home_category_3_amount,
+            ]
+
+            yield put(globalActions.receiveHomeScreenCategories(homeScreenCategories))
+            yield put(globalActions.receiveHomeScreenCategoryAmounts(homeScreenCategoryAmounts))
+
+            yield fork(getHomeScreenArticles)
         }
     } catch (err) {
         console.log('error in get custom domain options saga', err)
-        // default options for theme
-        yield put(
-            themeActions.saveTheme({
-                theme: 'light',
-                primary: '#2099CE',
-                accent: '#83B33B',
-            })
-        )
+
         yield put(globalActions.receiveHeader(''))
         yield put(globalActions.receiveHeaderLogo(''))
+        yield put(globalActions.receiveCommentsOption(false))
         yield put(globalActions.receiveSportCenterOption(false))
         yield put(globalActions.receiveStoryListStyle('small'))
+        yield put(globalActions.receiveHomeScreenListStyle('small'))
     }
 }
 
@@ -362,38 +291,6 @@ function* checkIfDomainIsInDb(domainId) {
     } catch (err) {
         console.log('error getting domain from DB', err)
         return []
-    }
-}
-
-function* initializeUser({ fromDeepLink }) {
-    try {
-        yield put(globalActions.initializeUserRequest())
-
-        yield call(findOrCreateUser)
-        yield call(loadActiveDomain)
-
-        yield put(globalActions.initializeUserSuccess())
-    } catch (err) {
-        console.log('error initializing user in saga', err)
-        yield put(globalActions.initializeUserError('error initializing user'))
-    }
-}
-
-function* initializeDeepLinkUser({ params: { schoolId } }) {
-    try {
-        // if deep link this will run
-        yield put(globalActions.initializeDeepLinkUserRequest())
-
-        console.log('in deep link init', schoolId)
-        yield call(findOrCreateUser)
-
-        SplashScreen.hide()
-        NavigationService.navigate('DeepSelect', { schoolId: schoolId })
-
-        yield put(globalActions.initializeDeepLinkUserSuccess())
-    } catch (err) {
-        console.log('error initializing deep link user in saga', err)
-        yield put(globalActions.initializeDeepLinkUserError('error initializing deep link user'))
     }
 }
 
@@ -439,7 +336,6 @@ function* startupSaga() {
     yield all([
         takeLatest(globalTypes.STARTUP, startup),
         takeLatest(globalTypes.INITIALIZE_USER, initializeUser),
-        takeLatest(globalTypes.INITIALIZE_DEEP_LINK_USER, initializeDeepLinkUser),
         takeLatest(globalTypes.FETCH_HOME_SCREEN_ARTICLES, getHomeScreenArticles),
     ])
 }
