@@ -1,521 +1,48 @@
-import React from 'react'
-import {
-    ScrollView,
-    StyleSheet,
-    View,
-    TextInput,
-    Text,
-    Image,
-    Alert,
-    ActivityIndicator,
-} from 'react-native'
+import React, { useState } from 'react'
+import { ScrollView, View, Text, Image, Alert, ActivityIndicator } from 'react-native'
 import * as Amplitude from 'expo-analytics-amplitude'
 import * as Haptics from 'expo-haptics'
 import * as WebBrowser from 'expo-web-browser'
 import Constants from 'expo-constants'
 
-import HTML from 'react-native-render-html'
+import { List, Divider, Switch, IconButton, Button, Portal } from 'react-native-paper'
 
-import { connect } from 'react-redux'
+import { Feather } from '@expo/vector-icons'
 
-import {
-    List,
-    Divider,
-    Switch,
-    IconButton,
-    Colors,
-    Snackbar,
-    Button,
-    Portal,
-} from 'react-native-paper'
+const SettingsScreen = (props) => {
+    const {
+        navigation,
+        userInfo,
+        domains,
+        theme,
+        deleteUser,
+        isLoading,
+        subscribe,
+        unsubscribe,
+        unsubscribeLoading,
+        deleteDomain,
+        removeSchoolSub,
+        setActiveDomain,
+        setInitialized,
+        toggleDarkMode,
+    } = props
 
-import { actions as domainActions, getActiveDomain } from '../redux/domains'
-import { types as userTypes, actions as userActions } from '../redux/user'
-import { actions as globalActions } from '../redux/global'
-import { createLoadingSelector } from '../redux/loading'
-import { createErrorMessageSelector } from '../redux/errors'
-
-const deleteUserLoadingSelector = createLoadingSelector([userTypes.DELETE_USER])
-const deleteUserErrorSelector = createErrorMessageSelector([userTypes.DELETE_USER])
-
-const unsubscribeLoadingSelector = createLoadingSelector([userTypes.UNSUBSCRIBE])
-
-const ActiveDomainIcon = ({ color }) => <List.Icon icon={`star`} color={color} />
-
-class SettingsScreen extends React.Component {
-    static navigationOptions = ({ navigation }) => {
-        const logo = navigation.getParam('headerLogo', null)
-        return {
-            title: 'Settings',
-            headerLeft: logo && (
-                <Image
-                    source={{ uri: logo }}
-                    style={{ width: 60, height: 35, borderRadius: 7, marginLeft: 10 }}
-                    resizeMode='contain'
-                />
-            ),
-            headerTitleAlign: 'center',
-        }
-    }
-
-    state = {
-        snackbarVisible: false,
-        editingUsername: false,
-        editingEmail: false,
-        username: '',
-        email: '',
-        notifications: {},
-        dialog: false,
-        clearedAllSettings: false,
-        showDeviceId: false,
-    }
-
-    componentDidMount() {
-        const { userInfo, domains, global, navigation } = this.props
-        this.setState({
-            username: userInfo.username,
-            email: userInfo.email,
-        })
-        navigation.setParams({
-            headerLogo: global.headerSmall,
-        })
-
-        this.setState({
-            notifications: domains.reduce(function (map, domain) {
-                map[domain.id] = domain.notificationCategories.reduce(function (map, notification) {
-                    map[notification.id] = notification.active
-                    return map
-                }, {})
+    const [dialog, setDialog] = useState(false)
+    const [notifications, setNotifications] = useState(
+        domains.reduce((map, domain) => {
+            map[domain.id] = domain.notificationCategories.reduce(function (map, notification) {
+                map[notification.id] = notification.active
                 return map
-            }, {}),
-        })
-    }
+            }, {})
+            return map
+        }, {})
+    )
 
-    render() {
-        const {
-            snackbarVisible,
-            editingUsername,
-            editingEmail,
-            username,
-            email,
-            notifications,
-        } = this.state
-        const {
-            domains,
-            userInfo,
-            theme,
-            errors,
-            deleteUser,
-            saveUserInfo,
-            isLoading,
-            unsubscribe,
-            unsubscribeLoading,
-        } = this.props
-
-        if (isLoading) {
-            return (
-                <View
-                    style={{
-                        flex: 1,
-                        justifyContent: 'center',
-                        alignItems: 'center',
-                    }}
-                >
-                    <ActivityIndicator />
-                </View>
-            )
-        }
-
-        return (
-            <ScrollView style={styles.container}>
-                <View style={{ flex: 1 }}>
-                    <List.Section>
-                        <List.Subheader>User Preferences</List.Subheader>
-                        {editingUsername ? (
-                            <TextInput
-                                style={{ height: 40, width: 250, fontSize: 15, paddingLeft: 60 }}
-                                onBlur={(text) => {
-                                    saveUserInfo({
-                                        username: username,
-                                        email: email,
-                                    })
-                                    this.setState({
-                                        editingUsername: false,
-                                    })
-                                }}
-                                autoFocus={true}
-                                returnKeyType='done'
-                                value={username}
-                                onChangeText={(text) => this.setState({ username: text })}
-                            />
-                        ) : (
-                            <List.Item
-                                title={userInfo.username || 'Not Set'}
-                                description={'Username'}
-                                style={styles.inactiveItem}
-                                right={() => {
-                                    return (
-                                        <IconButton
-                                            icon='plus'
-                                            color={Colors.grey300}
-                                            size={28}
-                                            onPress={() => this._handleUserInfoEdit('username')}
-                                        />
-                                    )
-                                }}
-                            />
-                        )}
-                        {editingEmail ? (
-                            <TextInput
-                                style={{ height: 40, width: 250, fontSize: 15, paddingLeft: 60 }}
-                                onBlur={() => {
-                                    saveUserInfo({
-                                        username: username,
-                                        email: email,
-                                    })
-                                    this.setState({
-                                        editingEmail: false,
-                                    })
-                                }}
-                                autoFocus={true}
-                                returnKeyType='done'
-                                value={email}
-                                onChangeText={(text) => this.setState({ email: text })}
-                            />
-                        ) : (
-                            <List.Item
-                                title={userInfo.email || 'Not Set'}
-                                description={'email'}
-                                style={styles.inactiveItem}
-                                right={() => {
-                                    return (
-                                        <IconButton
-                                            icon='plus'
-                                            color={Colors.grey300}
-                                            size={28}
-                                            onPress={() => this._handleUserInfoEdit('email')}
-                                        />
-                                    )
-                                }}
-                            />
-                        )}
-                    </List.Section>
-                    <Divider />
-                    <List.Section>
-                        <List.Subheader>Saved Organizations</List.Subheader>
-                        {domains.map((item) => {
-                            return (
-                                <List.Item
-                                    key={item.id}
-                                    title={item.name}
-                                    style={!item.active ? styles.inactiveItem : null}
-                                    description={item.active ? 'active' : null}
-                                    left={() => {
-                                        if (item.active) {
-                                            return <ActiveDomainIcon color={theme.colors.accent} />
-                                        } else {
-                                            return null
-                                        }
-                                    }}
-                                    right={() => {
-                                        return (
-                                            <IconButton
-                                                icon='delete'
-                                                color={Colors.red700}
-                                                size={28}
-                                                onPress={() => this._handleDeleteOrg(item)}
-                                            />
-                                        )
-                                    }}
-                                    onPress={() => {
-                                        this._switchDomain(item.id)
-                                    }}
-                                />
-                            )
-                        })}
-                    </List.Section>
-                    <Divider />
-                    <List.Item
-                        style={{ paddingVertical: 0 }}
-                        title='Add New Organization'
-                        left={() => <List.Icon icon={`plus`} />}
-                        onPress={this._handleAddNewOrg}
-                    />
-                    <Divider />
-                    <List.Section>
-                        <List.Subheader>Push Notifications</List.Subheader>
-                        {userInfo.user.push_token ? (
-                            domains.map((domain) => {
-                                const writerSubs = userInfo.writerSubscriptions.filter(
-                                    (writer) => writer.organization_id === domain.id
-                                )
-                                return (
-                                    <List.Accordion
-                                        key={domain.id}
-                                        title={domain.name}
-                                        left={(props) => (
-                                            <List.Icon {...props} icon='folder-open' />
-                                        )}
-                                    >
-                                        <List.Subheader>Writer Notifications</List.Subheader>
-                                        {writerSubs.length > 0 ? (
-                                            writerSubs.map((writerObj) => {
-                                                return (
-                                                    <List.Item
-                                                        key={writerObj.id}
-                                                        style={{
-                                                            paddingVertical: 0,
-                                                            paddingLeft: 60,
-                                                        }}
-                                                        title={writerObj.writer_name}
-                                                        right={() => {
-                                                            return unsubscribeLoading ? (
-                                                                <ActivityIndicator
-                                                                    style={{ paddingRight: 10 }}
-                                                                />
-                                                            ) : (
-                                                                <IconButton
-                                                                    icon='delete'
-                                                                    color={Colors.red700}
-                                                                    size={20}
-                                                                    onPress={() =>
-                                                                        unsubscribe({
-                                                                            subscriptionType:
-                                                                                'writers',
-                                                                            ids: [writerObj.id],
-                                                                            domainId: domain.id,
-                                                                        })
-                                                                    }
-                                                                />
-                                                            )
-                                                        }}
-                                                    />
-                                                )
-                                            })
-                                        ) : (
-                                            <Text
-                                                style={{
-                                                    fontSize: 18,
-                                                    fontWeight: 'bold',
-                                                    paddingBottom: 10,
-                                                }}
-                                            >
-                                                You aren't following any writers yet
-                                            </Text>
-                                        )}
-
-                                        <List.Subheader>Category Notifications</List.Subheader>
-                                        {domain.notificationCategories.map((item, i) => {
-                                            if (item.category_name == 'custom_push') {
-                                                return (
-                                                    <List.Item
-                                                        key={item.id}
-                                                        style={{
-                                                            paddingVertical: 0,
-                                                            paddingLeft: 60,
-                                                        }}
-                                                        title='Alerts'
-                                                        right={() => {
-                                                            return (
-                                                                <Switch
-                                                                    style={{ margin: 10 }}
-                                                                    value={
-                                                                        notifications[domain.id][
-                                                                            item.id
-                                                                        ]
-                                                                    }
-                                                                    onValueChange={(value) => {
-                                                                        this._toggleNotifications(
-                                                                            item.id,
-                                                                            value,
-                                                                            domain,
-                                                                            item
-                                                                        )
-                                                                    }}
-                                                                />
-                                                            )
-                                                        }}
-                                                    />
-                                                )
-                                            }
-                                            return (
-                                                <List.Item
-                                                    key={item.id}
-                                                    style={{ paddingVertical: 0, paddingLeft: 60 }}
-                                                    title={
-                                                        <HTML
-                                                            html={item.category_name}
-                                                            customWrapper={(text) => {
-                                                                return (
-                                                                    <Text
-                                                                        ellipsizeMode='tail'
-                                                                        numberOfLines={1}
-                                                                        style={{ fontSize: 16 }}
-                                                                    >
-                                                                        {text}
-                                                                    </Text>
-                                                                )
-                                                            }}
-                                                            baseFontStyle={{ fontSize: 16 }}
-                                                        />
-                                                    }
-                                                    right={() => {
-                                                        return (
-                                                            <Switch
-                                                                style={{ margin: 10 }}
-                                                                value={
-                                                                    notifications[domain.id][
-                                                                        item.id
-                                                                    ]
-                                                                }
-                                                                onValueChange={(value) => {
-                                                                    this._toggleNotifications(
-                                                                        item.id,
-                                                                        value,
-                                                                        domain,
-                                                                        item
-                                                                    )
-                                                                }}
-                                                            />
-                                                        )
-                                                    }}
-                                                />
-                                            )
-                                        })}
-                                    </List.Accordion>
-                                )
-                            })
-                        ) : (
-                            <Text
-                                style={{
-                                    textAlign: 'center',
-                                    fontSize: 18,
-                                    fontWeight: 'bold',
-                                    paddingBottom: 10,
-                                }}
-                            >
-                                You have disabled push notifications for this app
-                            </Text>
-                        )}
-                    </List.Section>
-                    <Divider />
-                    <List.Item
-                        title='Privacy Policy'
-                        style={{ paddingVertical: 0 }}
-                        left={(props) => <List.Icon {...props} icon='account-supervisor' />}
-                        onPress={this._viewLink}
-                    />
-                    <View>
-                        <Button
-                            icon='delete-forever'
-                            mode='outlined'
-                            color={Colors.red700}
-                            style={{ padding: 10 }}
-                            onPress={() => {
-                                this.setState({
-                                    dialog: true,
-                                    showDeviceId: !this.state.showDeviceId,
-                                })
-                            }}
-                        >
-                            Clear All Settings
-                        </Button>
-                    </View>
-                </View>
-                <Snackbar
-                    visible={snackbarVisible}
-                    duration={3000}
-                    style={{
-                        position: 'absolute',
-                        bottom: 0,
-                        left: 0,
-                        right: 0,
-                    }}
-                    onDismiss={() => this.setState({ snackbarVisible: false })}
-                    action={{
-                        label: 'Dismiss',
-                        onPress: () => {
-                            this.setState({ snackbarVisible: false })
-                        },
-                    }}
-                >
-                    Organization Removed
-                </Snackbar>
-                <Snackbar
-                    visible={errors}
-                    duration={3000}
-                    style={{
-                        position: 'absolute',
-                        bottom: 100,
-                        left: 0,
-                        right: 0,
-                    }}
-                    onDismiss={() => {}}
-                    action={{
-                        label: 'Dismiss',
-                        onPress: () => {},
-                    }}
-                >
-                    Sorry there was an error clearing your settings. Please try again later.
-                </Snackbar>
-                <Portal>
-                    {this.state.dialog
-                        ? Alert.alert(
-                              'Clear all settings?',
-                              'This will clear all of your saved schools, articles, and notification settings.',
-                              [
-                                  {
-                                      text: 'Cancel',
-                                      onPress: this._hideDialog,
-                                      style: 'cancel',
-                                  },
-                                  {
-                                      text: 'Clear',
-                                      onPress: () => {
-                                          deleteUser()
-                                          this._hideDialog()
-                                      },
-                                  },
-                              ],
-                              { cancelable: false }
-                          )
-                        : null}
-                </Portal>
-                {this.state.showDeviceId && (
-                    <Text
-                        style={{
-                            alignSelf: 'flex-end',
-                            marginTop: 'auto',
-                            fontSize: 9,
-                            color: Colors.grey400,
-                        }}
-                    >
-                        {`${Constants.manifest.version} - ${Constants.installationId}`}
-                    </Text>
-                )}
-            </ScrollView>
-        )
-    }
-
-    _hideDialog = () => this.setState({ dialog: false })
-
-    _viewLink = async () => {
-        let result = await WebBrowser.openBrowserAsync('https://snosites.com/privacy-policy/')
-    }
-
-    _handleAddNewOrg = () => {
+    const _handleDeleteOrg = (domain) => {
         Haptics.selectionAsync()
-        this.props.navigation.navigate('Auth')
-    }
-
-    _handleDeleteOrg = (domain) => {
-        Haptics.selectionAsync()
-        const { domains, navigation, deleteDomain, unsubscribe, removeSchoolSub } = this.props
 
         // get all the category IDs to remove them from DB
-        const categoryIds = domain.notificationCategories.map((category) => {
-            return category.id
-        })
+        const categoryIds = domain.notificationCategories.map((category) => category.id)
         if (categoryIds) {
             unsubscribe({
                 subscriptionType: 'categories',
@@ -524,14 +51,10 @@ class SettingsScreen extends React.Component {
             })
         }
 
-        this.setState({
-            snackbarVisible: true,
-        })
-        // send analytics data
+        // send old analytics data
         Amplitude.logEventWithProperties('remove school', {
             domainId: domain.id,
         })
-
         //new analytics
         removeSchoolSub(domain.url)
 
@@ -539,42 +62,30 @@ class SettingsScreen extends React.Component {
             let found = domains.find((domain) => {
                 return !domain.active
             })
+            Haptics.selectionAsync()
             if (found) {
-                this._switchDomain(found.id)
+                setActiveDomain(found.id)
+                setInitialized(false)
             } else {
-                navigation.navigate('AuthLoading')
+                setActiveDomain(null)
+                setInitialized(false)
             }
         }
 
         deleteDomain(domain.id)
     }
 
-    _handleUserInfoEdit = (userPref) => {
-        if (userPref === 'username') {
-            this.setState({
-                editingUsername: true,
-            })
-        } else if (userPref === 'email') {
-            this.setState({
-                editingEmail: true,
-            })
-        }
-    }
-
-    _toggleNotifications = (notificationId, value, domain, notification) => {
+    const _toggleNotifications = (notificationId, value, domain, notification) => {
         Haptics.selectionAsync()
         // stops lag of DB call for switch value
-        this.setState({
-            notifications: {
-                ...this.state.notifications,
-                [domain.id]: {
-                    ...this.state.notifications[domain.id],
-                    [notificationId]: value,
-                },
+        setNotifications({
+            ...notifications,
+            [domain.id]: {
+                ...notifications[domain.id],
+                [notificationId]: value,
             },
         })
 
-        const { subscribe, unsubscribe } = this.props
         if (value) {
             subscribe({
                 subscriptionType: 'categories',
@@ -590,49 +101,190 @@ class SettingsScreen extends React.Component {
         }
     }
 
-    _switchDomain = (id) => {
-        Haptics.selectionAsync()
-        const { changeActiveDomain, navigation } = this.props
-
-        changeActiveDomain(id)
-        navigation.navigate('AuthLoading')
+    if (isLoading) {
+        return (
+            <View
+                style={{
+                    flex: 1,
+                    justifyContent: 'center',
+                    alignItems: 'center',
+                }}
+            >
+                <ActivityIndicator />
+            </View>
+        )
     }
+    return (
+        <ScrollView
+            style={{
+                flex: 1,
+                backgroundColor: theme.colors.surface,
+            }}
+        >
+            <View style={{ flex: 1 }}>
+                <List.Section>
+                    <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}>
+                        <List.Subheader>User Preferences</List.Subheader>
+                        <IconButton
+                            style={{ marginRight: 20 }}
+                            icon='wrench'
+                            color={theme.colors.primary}
+                            size={20}
+                            onPress={() => navigation.push('UserInfoModal')}
+                        ></IconButton>
+                    </View>
+                    <List.Item
+                        title={userInfo.username || 'Not Set'}
+                        description={'Username'}
+                        style={{
+                            paddingLeft: 60,
+                        }}
+                    />
+                    <List.Item
+                        title={userInfo.email || 'Not Set'}
+                        description={'email'}
+                        style={{
+                            paddingLeft: 60,
+                        }}
+                    />
+                </List.Section>
+                <Divider />
+                <List.Section>
+                    <List.Subheader>Saved Organizations</List.Subheader>
+                    {domains.map((item) => {
+                        return (
+                            <List.Item
+                                key={item.id}
+                                title={item.name}
+                                style={!item.active ? { paddingLeft: 60 } : null}
+                                description={item.active ? 'active' : null}
+                                left={() =>
+                                    item.active ? (
+                                        <List.Icon icon={`star`} color={theme.colors.accent} />
+                                    ) : null
+                                }
+                                right={() => (
+                                    <IconButton
+                                        icon='delete'
+                                        color={theme.colors.red}
+                                        size={28}
+                                        onPress={() => _handleDeleteOrg(item)}
+                                    />
+                                )}
+                                onPress={() => {
+                                    Haptics.selectionAsync()
+                                    setActiveDomain(item.id)
+                                    setInitialized(false)
+                                }}
+                            />
+                        )
+                    })}
+                    <Button
+                        icon='plus'
+                        mode='text'
+                        contentStyle={{ padding: 10 }}
+                        onPress={() => {
+                            Haptics.selectionAsync()
+                            setActiveDomain(null)
+                        }}
+                    >
+                        Add New Organization
+                    </Button>
+                </List.Section>
+                <Divider />
+                <List.Subheader>Extras</List.Subheader>
+                <List.Item
+                    title='Privacy Policy'
+                    style={{ paddingVertical: 10 }}
+                    left={(props) => <List.Icon {...props} icon='account-supervisor' />}
+                    onPress={() =>
+                        WebBrowser.openBrowserAsync('https://snosites.com/privacy-policy/')
+                    }
+                />
+                <List.Item
+                    title='Terms of Service'
+                    style={{ paddingVertical: 10 }}
+                    left={(props) => <List.Icon {...props} icon='library-books' />}
+                    onPress={() =>
+                        WebBrowser.openBrowserAsync('https://snosites.com/terms-of-service/')
+                    }
+                />
+                <List.Item
+                    title={'Device ID'}
+                    description={Constants.installationId}
+                    descriptionStyle={{ fontSize: 12 }}
+                    style={{ paddingVertical: 10 }}
+                    left={(props) => <List.Icon {...props} icon='cellphone-information' />}
+                />
+                <List.Item
+                    title='Clear All Settings'
+                    titleStyle={{ color: theme.colors.red }}
+                    style={{ paddingVertical: 10 }}
+                    left={(props) => (
+                        <List.Icon {...props} icon='delete-forever' color={theme.colors.red} />
+                    )}
+                    onPress={() => {
+                        setDialog(true)
+                    }}
+                />
+                <Divider />
+                <View
+                    style={{
+                        flexDirection: 'row',
+                        justifyContent: 'space-between',
+                        padding: 20,
+                        alignItems: 'center',
+                    }}
+                >
+                    <Text
+                        style={{
+                            fontFamily: 'raleway',
+                            textAlign: 'center',
+                            color: theme.colors.text,
+                        }}
+                    >
+                        The Source version: {Constants.manifest.version}
+                    </Text>
+                    <View style={{ flexDirection: 'row' }}>
+                        <Switch
+                            value={theme.dark}
+                            onValueChange={(value) => toggleDarkMode(value)}
+                            color={theme.colors.primary}
+                        />
+                        <Feather
+                            style={{ marginLeft: 20, alignSelf: 'center' }}
+                            name={theme.dark ? 'moon' : 'sun'}
+                            size={23}
+                            color={theme.dark ? 'white' : '#ebbb0e'}
+                        />
+                    </View>
+                </View>
+            </View>
+            <Portal>
+                {dialog
+                    ? Alert.alert(
+                          'Clear all settings?',
+                          'This will clear all of your saved schools, articles, and notification settings.',
+                          [
+                              {
+                                  text: 'Cancel',
+                                  onPress: () => setDialog(false),
+                                  style: 'cancel',
+                              },
+                              {
+                                  text: 'Clear',
+                                  onPress: () => {
+                                      deleteUser()
+                                      setDialog(false)
+                                  },
+                              },
+                          ],
+                          { cancelable: false }
+                      )
+                    : null}
+            </Portal>
+        </ScrollView>
+    )
 }
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#fff',
-        // justifyContent: 'space-between',
-    },
-    inactiveItem: {
-        paddingLeft: 60,
-    },
-})
-
-const mapStateToProps = (state) => {
-    return {
-        theme: state.theme,
-        domains: state.domains,
-        userInfo: state.user,
-        global: state.global,
-        activeDomain: getActiveDomain(state),
-        errors: deleteUserErrorSelector(state),
-        isLoading: deleteUserLoadingSelector(state),
-        unsubscribeLoading: unsubscribeLoadingSelector(state),
-    }
-}
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-        changeActiveDomain: (domainId) => dispatch(domainActions.setActiveDomain(domainId)),
-        deleteUser: () => dispatch(userActions.deleteUser()),
-        saveUserInfo: (payload) => dispatch(userActions.saveUserInfo(payload)),
-        deleteDomain: (domainId) => dispatch(domainActions.deleteDomain(domainId)),
-        subscribe: (payload) => dispatch(userActions.subscribe(payload)),
-        unsubscribe: (payload) => dispatch(userActions.unsubscribe(payload)),
-        removeSchoolSub: (url) => dispatch(globalActions.removeSchoolSub(url)),
-    }
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(SettingsScreen)
+export default SettingsScreen
